Use the vue-router to reflect the query parameters of the iframe in the outer url.
Why: this way a page reload will just move again to the current view
instead of just loading the start page. We also push a new entry to the
history stack when we get an iframe load event. This way the user can
just navigate as if it would open Redaxo as stand-alone app.